chore add ignore type for gorm.DeletedAt at BrokerEntity

+++ src/services/inventory/contracts/Paginated.go
@@ -1,8 +1,8 @@
 package contracts
 
 type PaginatedResult[T any] struct {
-	Result     []*T `json:"result"`
-	PageNumber int  `json:"pageNumber"`
-	PageSize   int  `json:"pageSize"`
-	TotalItems int  `json:"totalItems"`
+	Result     []*T  `json:"result"`
+	PageNumber int   `json:"pageNumber"`
+	PageSize   int   `json:"pageSize"`
+	TotalItems int64 `json:"totalItems"`
 }
